Refactor connection code
